This patch adds a flag '-c' to the radvd binary which will exit radvd once completing the startup tests. This is similar to the 'configtest' option in apache's apachectl command. This patch is built against the current code available via cvs.
I coded this patch as I am working on a programmatic way of adding and
removing subnets from a linux router. For health paranoia's sake I
wanted to check the new configuration that is generated is valid for
radvd to run against.

[Second patch merged, explanation:]
It does make it a simpler patch to use flog(). I originally didn't use
it because I wanted to make sure the output went to stderr regardless
of other command line options. By overriding the value of 'log_method'
if configtest is called the existing logging code can be used with no
changes.

From Iain J. Watson.