process netlink messages and reload_config
This is still bad.  If any of our interfaces gets a netlink message, we
reload the config.  Reloading the config causes radvd to almost start
over.  A netlink message on one interface will affect all the other
interfaces.  There must be a better way to do this, but that's for
a later commit.