Correctly handle setup_linklocal_addr failure
If an interface with IgnoreIfMissing doesn't have a link-local address during
reconfiguration, not only the the address, but also the interface index is
missing from the Interface structure. As address configuration changes don't
trigger a config reload, this could lead to a permanent failure when there was
nothing else to trigger a reload.