time: Use clock_gettime and monotonic clock
Currently the system time is used to calculate the timeout in
milliseconds for the poll function. This can lead to bugs as soon as the
time of the system changes through NTP. For example starting radvd
before NTP can lead to a 44 year jump, resulting in a poll timeout of
10^6 seconds.

This patch replaces all gettimeofday() usage by the better clock_gettime
call using a monotonic clock which avoids those situations.

Signed-off-by: Markus Pargmann <mpa@pengutronix.de>