send.c: make const what we can make const
Some of the add functions are modifying the iface in a way I don't
like.  The decr lifetime features is one of the reasons.  Another
is an update to the last time an interface was used.  I'd like to
factor out the code that changes these so 'sending' on and interface
doesn't also *change* an interface.