recv.c: allocate mhdr.msg_control in the callers
recv_rs_ra was calling malloc and setting a static buffer for this
field.  The memory leak always bothered me.  This change uses
a stack buffer in the callers of recv_rs_ra for this value.

The parameter list is getting a bit long and complicated for this
function, so there's more to do to clean this up.