privsep-linux.c: umask(0), setsid(), chdir("/")
privsep_init should ensure its umake is zero, it's a session/group leader
and to change dir to '/' to act like a proper daemon.  I'd like the
privsep process to be in the same session/group as the radvd child process,
but I always was to fork off privsep before the config file is read.
Forking after reading the config file uses extra memory in privsep for no
good reason.