comment out all disable_ipv6_autoconf code
This code is causing a problem in a setup with multiple routers
running multiple radvd on a network.  One of the radvd machines
isn't getting the routes from the other because autoconf is
completely disabled by RADVD on the interface it advertises.  The
motivation to disable autoconf was to prevent an interface from
autoconfiguring using radvd's on advertisements, so it was mostly
a cosmetic thing anyway.