move the packages check after creating packages
This check is here to prevent radvd from being released without
the GIT tree being tagged.  There are still sometime cases where
a packages can be generated for testing, so check if the tree is
tagged after the packages are created.  It will still fail the build,
but the packages will be usable.