remove set autoconf=0 code
This code was commented out and tested by mikken.  I suspect this
code was also responsible for Nicolas's problem on FreeBSD where
radvd wouldn't setup any interfaces.  set_autoconf always failed
on FreeBSD which caused setup_iface to always fail.

This code was added to prevent interfaces on which radvd advertises
to not use their own adverts to configure addresses.  Since this
isn't technically necessary, and because it's causing problems, I'm
removing it completely.