radvd.c: fix SIGHUP mem leak
reload config frees and allocates a new interface list, but only
changes the pointer to it in main_loop.  main still had a pointer
to the freed list and would attempt to free it before exiting. The
trick is to pass the pointer from main_loop back to main as a return
value to main will free the correct interface list.