process.c: make sure we wait for random delay
When receiving an RS, we need to wait a random amount of time before
sending out our RA to avoid packet collision in the lower layers of
the network stack.  As far as I know, the lower levels will deal with
the collision, but the RADVD RFC calls for a random delay so this
change is making making sure it happens.