maintain address list on each iface
Maintain a sorted list of addresses on each iface.  The list is going to be
used by the netlink code to ensure that an RTM_(NEW|DEL)ADDR message is not a
false alarm.  It's sorted to ensure that the order the OS returns addresses
doesn't matter and it will be trivially easy to identify new or deleted
addresses.