check bogusness of RTM_(NEW|DEL)ADDR messages
When NetworkManager is running, radvd gets RTM_NEWADDR messages on
the netlink socket each time radvd sends out an RA.  This change adds
a check in netlink to campare the previous list of addresses on an
iface to the current list after the RTM_NEWADDR message is received.
If the list is the same, the RTM_NEWADDR can be ignored.