Check lifetime of explicit 0::/0 vs AdvDefaultLifetime
For the default route 0::/0, we need to explicitly check the lifetime
against the AdvDefaultLifetime value.

If exactly one of the two has a zero value, then nodes processing the RA
may have a flap in their default route (confirmed on Linux 3.13):
AdvDefaultLifetime == 0 && route.AdvRouteLifetime > 0:
- default route is deleted and then re-added.
AdvDefaultLifetime > 0 && route.AdvRouteLifetime == 0:
- default route is added and then deleted.

Issue a warning for this case, but do not abort to avoid breaking
existing configurations.

Fixes: https://github.com/reubenhwk/radvd/issues/65
Signed-off-by: Robin H. Johnson <robbat2@gentoo.org>