test/send: replace print_safe_buffer with snprint_safe_buffer & assert.
Previously the code just printed out all of the buffers in whatever
order the tests ran in, which made it hard to trace which printed buffer
belonged to a given test.

Use snprint_safe_buffer instead to print to a string, and then assert a
failure with the string; gives nice prefixed output about where it goes.

Signed-off-by: Robin H. Johnson <robbat2@gentoo.org>