AdvRASolicitedUnicast: unicast RA response to RS.
RFC 7772 section 2.1 describes a network flooded with multicast RA at
the maximum 3 second interval due a high client turnover.

The mitigation described in RFC 7772 section 5.1.1 states that a unicast
RA response can be sent, but SHOULD be a configurable option
(AdvRASolicitedUnicast), and that networks containing tens or hundreds
of battery-powered devices SHOULD enable the option.

The new option is defaulted to on, as it has very few downsides, and
represents significent battery life improvements for many clients.

We do differ from the RFC in that we do not require the SLLA option to
be set in the RS, as testing shows many clients are not setting it.

See TODO for further possible improvements to AdvRASolicitedUnicast.

Fixes: https://github.com/reubenhwk/radvd/issues/63
Fixes: https://github.com/reubenhwk/radvd/pull/69
Signed-off-by: Robin H. Johnson <robbat2@gentoo.org>