change handling of default log method
The default behavior is to daemonize, however the default log level is
L_STDERR_SYSLOG which leaves STDERR bleeding into the console. Thus,
change the default log method to L_SYSLOG (which fits daemonizing) and
automatically change it to L_STDERR_SYSLOG if --nodaemon was specified
on the command-line. However, the latter only takes effect, if the log
method wasn't specified explicitly.

Fixes #71.