Added contrib/95-radvd-gen.
A script that will generate radvd.conf from a template with dynamic
prefix lifetimes generated from the network interfaces (using the linux
ip command).  The script will check if the timing are correct in the
current radvd.conf and either:

a) generate a new radvd.conf and SIGHUP radvd, or
b) SIGUSR1 radvd to have the advertised timing reset, if required.

This is useful as a NetworkManager dispatcher script to keep the
timings in sync with those received from dhcpd, but can be used in
any environment where it can be run periodically.

Signed-off-by: Scott Shambarger <devel@shambarger.net>