No router advertisements on tunnel interfaces
The upstream commit 0c4dfeb889c9 ("improve multicast check, remove
broadcast check") forces UnicodeOnly if an interface does not support
IFF_MULTICAST. This stops periodic RAs being sent over tunnels, which
previously used to work. The solution is to allow sending such
multicast packets over point-to-point links. This is acceptable for
directly connected devices, or for tunnels, the packets are
encapsulated and so look like unicast packets to intervening routers.

Signed-off-by: Mike Manning <mmanning@brocade.com>