fix wrong assignment of struct msghdr
depending on the architecure the structure of msghdr may have a
different alignment using padding fields. this will lead to a wrong
assignment of values of the field name isnt specified in initalization.

example of struct msghdr in x86_64 (musl)

struct msghdr {
        void *msg_name;
        socklen_t msg_namelen;
        struct iovec *msg_iov;
        int msg_iovlen, __pad1;
        void *msg_control;
        socklen_t msg_controllen, __pad2;
        int msg_flags;
};

in the current code __pad1 gets assigned instead of msg_control and so
on.

Signed-off-by: Sebastian Gottschall <s.gottschall@dd-wrt.com>