Fix case of AdvValidLifetime option
In the manual page, the NAT64 Lifetime option is presented as
AdvValidLifeTime, while the only accepted input is with lower case T:
AdvValidLifetime.

This fixes the documentation. Please note that there is still some
inconsistency, as ABRO still uses LifeTime with capital T and have it
recognized as a separate token.

Signed-off-by: Ondřej Caletka <ondrej@caletka.cz>