Add missing configure check for AM_PROG_AR
The automake build generates a static convenience library
(libradvd_parser.a) but the configure script does not include a check
for the 'ar' program used to generate that library. Without any check,
automake ends up using the tool 'ar' as a fallback.

This fails in some distro packaging or cross-compiling environments
where a specific prefixed 'ar' tool is needed.

Adding the AM_PROG_AR makes the configure script autodetect the prefixed
ar tool in a way that's compatible with normal distro packaging and
cross compiling tooling.